Fixed incorrect prescaler value
The reason why this patch didn't work initially was purely because of the timer2 prescaler, which didn't roughly match the frequency of RTC!

This commit patches TIMER2 with a prescaler value which triggers the timer at an interval of around 31.25us, which is as precise as I could possibly make it.

This results in roughly correct waves, although the signal is inverted.

The wave is still jittery, the reference manual suggest this is due to the innaccuracy of RTC events...