microbit: PortOut/PortIn now respects bit mask
The mbed PortOut and PortIn abstractions require the definition of a mask that
specifies which pins in the given port are to be used.

The nrf51822 implementation did not respect this funcitonality, and would
overwrite all pins defined as an output, regardless of whether they were
defined as part of the mask.