microbit: Patch to HF Ticker implementaiton to reduce glitching on wait_ms()
Updated ticker implementation to:

 - Ensure initialisation of overflowount variable to zero (was previously undefined)

  - Update internal counter 'read' operation to handle overflow conditions
differently. Instead of re-implementing the overflow handler in the 'read'
operation, it now leaves this to a single code path in the ISR. Instead, the
code is simply made aware of overflows and handles this as a local event. This
prevents the possibility of duplicate increments to overflowCount (which
appeared to occur under testing, especially when the mbed wait_ms busy-wait function is used).